update tx list
