+++ app/src/main/java/com/iwallic/app/pages/asset/AssetDetailActivity.kt
@@ -60,7 +60,6 @@ class AssetDetailActivity : BaseActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_asset_detail)
-        window.addFlags(WindowManager.LayoutParams.FLAG_TRANSLUCENT_STATUS)
         initParams()
         initDOM()
         initGAS()
