+++ app/src/main/res/layout/fragment_tx_confirmed.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="utf-8"?>
+<com.scwang.smartrefresh.layout.SmartRefreshLayout android:id="@+id/transaction_pager"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    app:srlEnableOverScrollDrag="true"
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+    <com.scwang.smartrefresh.layout.header.ClassicsHeader
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        app:srlAccentColor="?colorPrimary"/>
+    <android.support.v7.widget.RecyclerView
+        android:id="@+id/transaction_list"
+        android:overScrollMode="never"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"/>
+    <com.scwang.smartrefresh.layout.footer.ClassicsFooter
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        app:srlAccentColor="?colorPrimary"/>
+</com.scwang.smartrefresh.layout.SmartRefreshLayout>
